Rewrite whole lib from es6+ to es5
There were a problem using this lib in environments with module bundler (eg webpack), after building code. Bundled code was mixed with 'nonbabelled' code from node_modules and it was causing errors in browser console. It could be done also with babeling code but to keep in project as less as possible dependencies, the better solution was to rewrite code to es5.